Move context-streaming operational tests into one mod
